+++ test/checks_test.go
@@ -1,7 +1,6 @@
 package test
 
 import (
-	"context"
 	"io/ioutil"
 	"path/filepath"
 	"runtime"
@@ -57,7 +56,8 @@ func TestChecks(t *testing.T) {
 		assert.NoError(t, err)
 		c, err := config.Parse([]byte("checks:\n  " + tc.check + ": danger"))
 		assert.NoError(t, err)
-		result, err := validator.ValidateController(context.Background(), &c, *workload)
+		var result validator.ControllerResult
+		result, err = validator.ValidateController(&c, *workload)
 		assert.NoError(t, err)
 		summary := result.GetSummary()
 		if tc.failure {